// src/app/interfaces/items.ts
import { Image } from './image';
export class Items {
  itemid: string;
  itemname: string;
  itemnumber: string;
  itemprice: number;
  itemquantity: string;
  itemdescription: string;
  itemcolor: string;
  itemcountry: string;
  itemheight: string;
  itemwidth: string;
  itemthickness: string;
  itemweight: string;
  itemimage: Image;
}